Begin to write code to put params into workspace re #8398
My original plan of putting the table into a POCO XML document
then putting that into the workspace like LoadParameters does
is not on, because LoadParameters does something dodgy, that
will sometime need fixing and should not be replicated.
So I'm forced to put the parameters in directly. This would lead
to better code, but I presently have no idea how to do it.

Signed-off-by: Karl Palmen <karl.palmen@stfc.ac.uk>