refs #8520. Placeholder for correction object.
We need a way to pass correction strategies into quick rather than having it rely on the IDF definitions contained within. Wrapping up the algorithm calls so that they can be partially applied outside of quick is the most sensible way to do this. The new system tests are currently passing prior to making the change of swapping this functionality in.