Refs #8475 Save and Laod dialogs no longer throw on cancel
Before, when cancel was pressed on a save or load dialog it would throw an IOexception saying "file not found".

This was because the previous call was retrieving a string and when cancel was pressed that string was null. The script would then attempt to open whatever was in that string, regardless of the user pressing open/save or cancel.

In order to make it safer the dialog now calls the QFileDialog.exec_() in order to check if the user pressed Open/Save or cancel, then acts accordingly.