Refs #8532. Save on close in place and new file menu items
The save-on-close behavior is in place but not fully wired up yet. It relys on a modified flag which currently isn't set by anything, but hard coding the value makes it work.
The autosave-on-fail behavior is also in place and not wired up.

The file menu has had a few tweaks as they were related to this ticket:
* "Save Table" has been renamed and relabled as "Save As..." to conform with a universally recognised command. It's shortcut has also changed to Ctrl-Alt-S (taking inspiration from Notepad++)
* "Load Table" has been renamed and relabled as "Open Table..." to conform with a universally recognised command. It's shortcut hasn't changed and is still Ctrl-O
* "Re-Load Table" had been renamed and relabled as "Reload from Disk" (taking inspiration from Notepad++) - it's shortcut hasn't changed and is still Ctrl-R
* New Command - "Save" overwrites the existing file if one has been loaded in, or will be the same as Save As if it's a new file. It's shortcut is the universally recognised Ctrl-S
* New Command - "Close Refl Gui" has been added as originally the only way to close the GUI was the control on the title bar, which is bad UI design. It takes on the same close command as any mantid sub-window: Ctrl-F4

The "Save" menu item, Save-on-close and Autosave-on-fail actualy share the same function, with a boolean flag separating the functionality. The default is save/save-on-close behavior, boolean true is passed to trigger the autosave behavior.

In order to do the save-on-close behavior I've had to re-implement QMainwindow in ISIS_Reflectonomy.py, overriding the closeEvent.

In order ot follow better pyqt practice, the existing signals have also had their connection statements changed to the new [object].[trigger].connect([slot/function]) form.