Refs #8532. Modified flag and autosave wired up
The autosave is wired up to the destructor, even if that may be a bad idea in python - this is purely becuse we've not been told any specific cases where the entrie GUI failed in the old version so it's the best option for now.

The modified flag has been wired up to the table's cellChanged signal and its made sure the triggers as when the window's closeEvent is called it will focus the process button to make sure the table fires events.

A loading flag as been introduced to prevent the modified flag from being tripped when loading a file.