Fix the theta/phi angles for the forward scattering detectors.
They all had the incorrect sign as the VMS detpos.dat file says that they
all belong in the mirror positions to each other. The absolute scattering
angle remains the same so the original fitting is unaffected. It does,
however, affect the gamma background calculation as this relies on knowing
the distance between a given foil & detector.
Refs #8378