Refs #8560. Functionality fixed, waiting for blocking tickets
The script now returns the data in the format we want, and does a check to make sure the user has entered only an integer to save checking for an eID that won't exist

The ui has had a minor tewk in that the central widget has had it's name changed and a sensible name given to the layout it hosts. The ListWidget has also been expanded slightly. Compared to the old functionality,(as of 2.6 when this wokred last) the list box will no logner resize when given data, scrollbars will be used instead.

This ticket is now wating for #8531, #8532 and #8549 to be pushed to master so i can merge them in and make sure that the functionality doesn't break.