Refs #8560. Fixed intrument selection after breaking in another ticket
When I changed the signal connections from the old C++ style to the new pyqt style notation i inadvertently broke the instrument selection combobox.

This was because the default "Activated" signal carried the index of the new item, rather than the new contents like it used to. The same method couldn't be kept as combobox would only return unicode strings, which the config manager didn't like as they weren't compatible with C++ std::string.

The solution was to a new class-member level list of instruments in all caps and use it to populate the listbox in the subclass (removing the hard coded values from the UI file) and set the config manager meaning that the values used would be python strings (compatible with C++ std::string)