Re #8613. Remove code that bails out early if no events in a bank.
We need the bank name to be added to the list even if it's empty as it's
checked for if the BankNames property has been set. There's already a check
that means that no LoadBankFromDiskTask is created for the bank if it's
empty (the original justification for adding this code in commit [36540217]).

I don't think this change has much other effect, but I'll be relying
heavily on the tests to confirm that given the horrific complexity of
the file.