Put Workspace2D spectra delete in OpenMP loop.
On MSVC this improves the time taken to delete large workspaces that have
had their memory allocated on multiple threads, i.e. running ConvertUnits
and breaking the X-data sharing. For example, a workspace
with ~70000 spectra & ~2600 bins (~4.3Gb) takes ~45 seconds to delete
after a ConvertUnits call. This drops to 1-2s if the deallocation of the
memory also happens within a multi-threaded loop.
Refs #8216