refs #8617. ReflectometryReductionOneAuto refined and tested.
New tests using a base class to avoid duplication. Base class ensures that signature and behaviour of facade algorithm is exactly the same as the base one. In addition, additional tests check that the values supplied to the wrapped algorithm upon execution are what we would expect from the IDF.