refs #8656. Remove region of interest.
The region of interest can now be fully provided via the ProcessingInstructions because the idea is just to sum the spectra across the range provided. The region of the direct beam is now optional since the user may not want to perform this normalisation. The multi detector and point detector processing now only differ by one optional step. This means that we can now perform transmission corrections on multi detector runs, which we couldn't do before.