Refs #8560. Added spinbox for RB search depth
I've added a spinbox for setting the depth of the search and made a few general improvments to the UI at the same time.

The widgets are now spread across two rows as they would have been cramped with the new widgets.

The Instrument combobox now has a label.

Buddies have been set, each label is now paired with a widget.

Tab order is now set properly.