Refs #8590 continuing work on integrating startlivedata
The user is now prompted for what Accumulation method they want to use for the live data.

However there is a problem where quick crashed after it tries to get the workspace defualts on the live data workspace. I will need to investigate this.

I plan on merging master into this branch in order to gain the benefit of some of the newer edits to quick, in order to see if they fix this.