Refs #8560. Added maximum to spinbox and improved complete message.
The cycle depth spinbox is now more inteligent, its max value is now set from the number of cycle journals logged in the main jounral file

The search complete message now contains a value for the ammount of results found. Searches that yeild nothing are still deemed complete as it searched the amount it was told to but nothing existed.

Refactored a few lines in the txt search method to explain cnstant values, and added a comment about a concatentaion line.