Refs #8763 UI tweaks made, nothing functional yet.
New menu has been added (as this is the simplest way to add shortcuts) but they've not been wired up to a function yet.

Polarisation corrections combobox has been moved.

Tooltips added to many controls, but not the table yet as that requires editign teh functionality code.

Spinbox default value cahnged to 5.

Added a few separators to the menus.