refs #8733. Intermediate changes made. Quick working.
Working well in Quick, but not via the GUI. This is because the GUI uses hard-coded names which make life really really hard! So i'm trying to work around this for now. Not complete yet.

The Parameter file does not get picked up for the newer POLREF IDF, so I am having to add a new one that will be picked up. I have a ticket open where we will completely replace the IDFs, at that point, I can delete this parameter file.