refs #8794 Replaced vector by std::pair structure
I am unsure about it, as the functional equivalent of this function for ConvertToMD methods would return more then 2 values. And some functional modifications of this method may also return more then 2 values. But after some thinking decide to make the change for the time being and then see if expansion is necessary.