Further changes to remove the axes not joining up at the corners
This one disables the axes backbones (these are the ones that don't join up)
and enables a canvas border.
There is one exception that all Colour bar axes will always show it's backbone - otherwise it looks horrible

Most of this is done through a change to the stored preferences using a one off change so people can still change the values later.

re #8809