Re #8637. Add capability to specify multiple runs.
Each run is treated as a single point in the scan (in the order they're
given), a scan_index entry is inserted and the runs are merged together.
They can then be treated as other scan runs. This allows scans from
non-ADARA-enabled instruments to use the functionality of this interface.

The loading algorithm had to be changed from LoadEventNexus to Load for
this. This then required changes around that code to catch if the user
tries to load the wrong kind of file.