refs #8804 Implemented Martyn's suggestions for ISpectum method
Though with reservations. Of course it looks better and more generic, but for vector I know that size is achieved in const time and accessing the elements through brackets is almost the fastest operation possible. Have no idea how back is implemented.

Other comments are work in progress -- have to discuss how actually this method is implemented for events.