re #8861: Fix the problem of identation in the python script
This python script is triggered when the python.directories changes.
There are identation problems with it.

The problem was introduced in this commit 8557da27854b141ed2d4215632c395f5cc15d936,
by auto-identation of editor.

The way to ensure it does not happen again is to allow the compiler
to concatenate strings as now.