Refs #8887 Fixed problem that was stopping auto-calc
dq/q calculation was the problem that stopped everything occuring.

The gui now attempts to calculate dq/q if not present, but if the logs it needs aren't there it stops and forces the user to enter one.

If it's given dq/q or can calcualte one itself, it has no problem calulating qmin and qmax if neccessary.