Refs #8906 - Use a scaled version of the same monitor.
Instead of passing the resulting monitor from SliceEvent to
NormalizeToMonitor, just use a scaled version of the same monitor that was
passed to SliceEvent.

These changes should also make the process of having monitors in a
different time regime easier, which is an issue we will be facing in the
future.