Refs #9041 Separated plot functionality
The new column is now home to the 'Plot' button and plotting is no longer done automatically when processing.

Plot buttons start off as disabled and are enabled when their row has been processed.

Pressing the plot button will rebin and plot the processing output like was done automatically before.

The are a couple of bugs still: There is now less progress indication when processing; and loading a new table, modifying the table, or clearing the table doesn't reset the plot buttons.