Refs #9041 Re-worked reload functionality
there were a couple of bugs pertaining to the current table and the type of the filename (as it was unicode), they've been fixed. Some instances where the loggere is given a variable have been fxied to convert to string to make sure unicode doesn't cause an error.

The reload functioanlity has also been changed to check if there is a loaded tabled before checking the mod flag