Refs #9043 Beginning connecting the Edit Menu
A clipboard object is now retrieved and stored when setting up the UI

Copy, Cut and Paste are wired up to basic handlers that simply print out at present in order to test their shortcuts.

Clear is implemented as expected, clearing text from highlighted cells when Edit->Clear or Del is pressed.