Refs #9043 Copy, Cut, Paste and clear all work
All of the edit menu now works as expected.

Copy/Paste functionality is compatible with MS Excel, and works in a similair way to Excels pasting too in regards to behavior when the user supplies a single cell target or a multi-cell target.

Selection mode has been changed to Contiguous to prevent problems occuring with multiple selections.