Refs #9042 Theta is now retrieved or calculated if not supplied
Quick had a problem where it required theta as the way it was retrieving it didn't work.

In the case where it doesn't have theta it'll first try and retrieve theta from the sample logs, and failing that, calculate it manually

The function now uses the _sample_ws loaded earlier to retrieve from. The problem stemmed from the way we've been changing how parameters to quick are provided, as run_ws was an old variable that had been forgotten about in the midst of these changes.