Refs #9152 Error message less misleading
Even though before the error was okay, it wasn't clear to someone who didn't know about rebinning, nor was it always correct like in the case of "first boundary, width, last boundary, width, last boundary" as there are an odd number of boundary values.

The error now states that the number of parameters must be odd as this is what it checks for in the validator, and would mean that the correct number of boundaries is present.