Refs #9170 Restructured Reflectometry Interface
The interface now follows the same pattern as the ORNL SANS interface for declaration. The ReflGui class is now the window itself, this means it recieves the close signal from the mantid main window and can override some things within itself.

Tidied up the connections, as there was no need for most of the handlers.

Renamed most of the methods to they follow python's style more closely.

Added more whitespace