Re #9087. Use EMPTY_DBL instead of NaN for default property values.
The NaN caused trouble with validation on RHEL6.

Note that psi had a guard to not write out a value if it is 'empty'.
This obviously harks back to when the default had been EMPTY_DBL in
the past, but in the current implementation it will always write out
something, and that something will be NaN for the default. I've
preserved that behaviour.