Refs #8960 ILL Cosmos and ANSTO now inherit from a new base class
AsciiPointBase is a new class that contains the base functionality for the ILL Cosmos and ANSTO save routines. It holds the data writing, value checking and property declaring functionality and if any extra functionality is required theres a couple of spots to add it as overridden methods.

Unused headers left over have been removed