Implemented general comparison operators for PoldiPeak
PoldiPeak::greaterThan and PoldiPeak::lessThan take two shared pointers
as argument and a pointer to a member function of PoldiPeak which takes
no arguments and returns an UncertainValue. boost::bind is used to tie
the supplied instances to these member functions and call them. The
results are then actually compared. These comparison functions can be
passed to std::sort, again with bind (see unit tests for demonstration).
This way, comparison operators need to be defined once and work for all
(numeric) properties of a PoldiPeak.