Define a set of GlobalFixtures for the unittests.
Currently these clear the AnalysisDataService & AlgorithmManager when
their tearDownWorld() methods are called, which ensures that all
workspaces placed in these singletons are cleared up before the end of
main.

If this it not done then the process hangs on MSVC while trying to
delete a Workspace2D, which contains an OpenMP loop in its destructor and
omp is not allowed to be called on Windows outside of main.

Refs #8216