Refs #8962 Fixes to loader
The loader had a problem where large groups of commas at the start of the row would cause the loader to become confused and skip a cell. Fixed so that it now looks if it was the first column or not, rather than looking at the place of the last found comma.

The stitch ID has been fixed to be an int to help the save algorithm