refs #9257. No need to specify the radius any more.
Part of the speedup is removing the PeakBackground strategy for the
simplier Hard threshold. We also have no need to use the radius. This is
nicer as we just let the algorithm figure out the integrated region
associated with a peak with minimal help (other than the threshold of
course).