Refs #9167 Transfer process changed
The way runs are transferred is now much mroe efficient.

First a regex will check if the text being transferred starts with one or more numbers followed by a colon. If so it uses the numbers. If it's not a simple case of extracting numbers, it'll then check the ADS and look at the sample logs. If it's not in the ADS it'll try and load it first

The functionality that refreshes the runs list after the ADS fetch option is changed has also been removed after a word with Owen,