refs #9360. Use hash cache pre-check.
We create a hash of each pair of labels to avoid processing duplicates. This makes a huge difference to performance. I'm seeeing a consistent x6 speedup on my osx build in debug mode with the IntegratePeaksUsingClustersPerformanceTest.

This pre-checking is particularly important where the Threshold for IntegratePeaksUsingClusters has been set low, because in that instance there will be many neighbouring clusters and therefore many pairs with the same label ids.