Refs #9361 - Default to standard-named IDF.
In the cases where we're loading IDF's without first loading in the data
(this happens for the system tests but not while using the interface under
normal circumstances), then default to the IDF with name
"INST_Definition.xml" rather than the latest IDF.

This is really just a quick fix so that we can make changes to the new
SANS2D IDF and param name that don't break the system tests.  The real
solution would be to hold off loading any IDF's at all without first
knowing which version of the instrument should be used -- this will be
done under #9361.