Refs #9371 Split usage of the presenter into a loaded and blank.
ReflMainViewPresenter has been made abstract and sublassed to ReflBlankMainViewPresenter (which creates a blank workspace with one row on the fly) and ReflLoadedMainViewPresenter (which does exactly what ReflMainViewPresenter used to do)

The test for ReflMainViewPresenter, has been moved to ReflLoadedMainViewPresenter and has had the fixes owen mentioned applied.

Rejigged the layout on the UI, in leiu of a menubar, there's a temperary row of buttons that will act the same until i can move the functionaity elsewhere.