Clear gcc 4.8 warning about non-trivial move assignment operator
C++11 dictates that move assignment operators will only be auto-generated
when a set of circumstances is met. One of these is the lack of a
user-defined destructor. Add virtual destructors to the Fit hierarchy
avoids the compiler generating trivial move-assignment ops and avoids
the warning in virtually inheriting subclasses, e.g Python
IPeakFunctionAdapter.cpp.
Refs #9421