Re #9442. Skip DiffSphereTest suite if running in Valgrind.
These tests fail if running under valgrind (see comment in commit itself).
Lean on a Valgrind macro to determine if we're running within Valgrind
and skip the tests if we are so that the rest of the tests can run
to completion (I'm running the CurveFittingTest executable directly so
it's running all the tests together).

The required header from valgrind has been placed into the CurveFitting/test
directory so we're not creating a dependency. If some platforms don't like
it I'll conditionally include/use it on linux only.