refs #9509. Single threaded CCL works faster.
Small adjustment to the code to use single threaded CCL. Results in a 4-5 times speed up. This is because we have n * small images being processed rather than one large one. The overhead of the parallel execution completely dominates the speed up acheived from the asynchronous thread execution in this case.

The next step will be to make the algorithm parallel with respect to the individual peaks in the peak list. This should result in a further speedup.