Re #8342. Add a run number getter to the live listener interface.
This allows a concrete live listener to return the current run number.
Previously, there was a hack in place that tried to fetch it out of
the extracted workspace, but this only worked for MatrixWorkspace (it
used the getRunNumber method). This way will work for any kind of
returned workspace.

Currently, listeners that do not have the capability of obtaining a
run number will always return zero. This also happens to the value
returned by the SNS listener when an instrument is not in a run (though
there's also the runStatus method that will indicate this).

Concrete listeners remain responsible for adding the run number to the
workspace logs, if appropriate.