Refs #8562 - Do a check for valid flood files before reducing.
This is based on Gesner's patch, which only checks on the GUI side of
things.  It would be better to do this in the Python layer as well, but
this will have to wait until a later date.  I'm sitting on a large amount
of code to do with User File parsing that will deal with file validation a
lot better, and this will someday make it in as part of #9367, but we have
so many other things to do that it's just not a priority at the moment.