Re #9542. Create a monitor workspace.
The monitor workspace is created in the initialization step at the same
time as the 'main' workspace. It is embedded inside the data buffer
workspace using the new functionality for that (setMonitorWorkspace).
It shares the instrument etc. with this main workspace and has the
detector IDs wired up properly.

On each call to extractData() a fresh, empty copy of this monitor workspace
is created and placed within the similarly fresh data buffer workspace.