Re #9543. Handle the addition of monitor workspace chunks.
This step explicitly adds monitor workspaces in addition to the main
data. This means that if the monitor workspace is still present after
the processing step (a user could extract it there if they wanted), it
will be available for the post-processing or be carried right through
to the output if there is no post-processing step.