Re #9545. Add setting of the monitor workspace pointer.
This is in addition to the existing way of an additional ADS entry with
"_monitors" suffix. It allows clients to use ExtractMonitorWorkspace on
the main output workspace and not need to worry about the name of the
workspace (it will work if LoadPreNexus was run as a subalgorithm, which
is not the case for old way).