Re #9543. Make sure monitor WS is kept if not preserving events.
Also add tests for this.

There's a call to ConvertToMatrixWorkspace if the option to preserve
events is off in a live data run. This propagates any monitor workspace
through this algorithm, though note that in the case of the SNS listener
at least a Rebin or something needs to be performed on the monitor
workspace to set the X boundaries.