Separate the Sphinx doctree directories by bulder type.
Each builder type now has a separate environment so that the order of
running the targets doesn't make a difference. Shared doctree directories
meant that running doctest first just in vanilla python wouldn't
generate the screenshots and running the html target wouldn't reread
the unchanged source files to generate the images. It forced doctest
to run in MantidPlot and do the screenshot generation.

Refs #9639