Re #9544. Filter events while the run was paused.
You might think that the DAS would not save events from when a run was
paused but no, that would be much to easy (for us). Therefore filter
them out if an ADARA-style pause log is present (i.e. this is SNS only).
There is a configuration property that can be set to load them as before.