Refs #9501 More Unit tests complete
The rest of the unit tests have been ported

ReflectomteryReductionOneAuto recieved a fix to prevent an uncaught out of bounds error. Now throws a different error after a check wiht a sensible error message

ReflectometryWorkflowBase recieved a fix to prevent name-less workspaces causing problems when compared with the default value.