refs #9906. Create and use virtual refl instrument.
Make a very simple reflectometry instrument in memory and attach a single monitor and a single detector spectrum to it. Avoids All IO related to both data and instrument, and can be used to create tiny reflectometry workspaces.

Makes a huge difference to test speed. Only applied to ReflectometryReductionOne at this stage.