Removed some old commented out code in Instrument.
Found this while updating the unit test for this ticket.

This code was commented out over two years ago. If it needed to be kept,
it would have been missed by now. Given the exception that immediately
follows gives the exact same information, we can afford to clean this
up.

Refs #9959