refs #10029. Combine operations.
The test code looked lean, I therefore profiled the test suite on RHEL6. The major issue seemed to be the number of binary operations occuring. I have reduced these through some simple rearrangement of the calculations and caching of intermediate values. On my OSX enviroment this is reporting a 50% speed up for the offending test (test_run_PA_unity) which is according to the RHEL6 builds a major bottleneck to the suite.