Implement most of Graph::loadFromProject
This is the vast majority of loadFromProject implemented.

Yet to be implemented:
 - <spectrogram>
 - <AxisFormula (N)>

Unable to be implemented currently:
 - <CurveLabels>
 - <MantidYErrors>
 - <SkipPoints>

This is because their parsing depends on their order relative to the
definition of curves in the project file. We cannot handle this
correctly because we do not discern the order of elements when parsing.
To resolve this, we will either have to make TSVSerialiser a lot more
sophisticated, or break backwards compatibility for a few small cases.

Refs #9970.