refs #9539. Fix bridge link unit tests.
The issue here is that new DisjointElements are being placed on the stack
as temporaries at the time that m_parent is set. When they go out of scope
the same address can be used in the next round of the loop, therefore all
m_parents get set to the same value.