Allow runsphinx_*.py scripts to be called by anyone
The driver script for running the sphinx process can now be called by a
user without any special requirements. The main use case will be
running a doctest individually.

It should be noted that even restricting the run to a single file with the
new -R option will actually still cause Sphinx to read every source file
at least once. This is baked in to Sphinx and cannot be avoided. However,
the config pickling does mean that all source files are only read once and
then after this only changed files are reread.
Refs #9639
Refs #