Refs #9371 Fix caching bug.
When there is only one row in the reflectometry gui, updating a value on
it does not cause the table view to be updated. This was a caching
related problem. This commit fixes that problem and optimises the
dataChanged signal, so that only items that have been updated need to be
fetched by the table view.