refs #10187. Always close group.
A few issues fixed here. Firstly. I found that the algorithm was failing
when thrown a very large group workspace. This turned out to be because we
needed to close the group within the doExec call. A group i always opened,
so it is necessary to close it.

The second thing done is to fix SaveNexus so that it will pass through
group workspaces intact to SaveNexusProcessed. Under the previous
mechanism, this would not have happened, because the generic Algorithm
group processing code would have exploded out the individual workspaces.