Alter approach to generating random point in sample/container.
For hollow shapes the old approach was very inefficient. The new method
chops up the sample into discrete boxes and first samples the boxes and
then inside a box. There is also a possibility of improving it further
by first removing the boxes that are in a hollow part of the object at
the caching step but this will be done later.
Refs #9842