Simplify SampleEnvironment to be a collection of Objects
This offers a massive performance improvement for repeated intersection
calculations as it removes he need to consider the absolute location
of the components. It is also more consistent with the way the sample
shape is dealt with as that does not catre about absolute location either.
The location is defined within the instrument.
Refs #9842