Fix race condition in ParameterMap
If threads are competing to try and both check and update the map then
they will hit the appropriate PARALLEL_CRITICAL sections within the
relevant methods. The bug was that the critical sections all had different
names and so OpenMP would not block on an ::add while a ::get is being
performed. The critical secctions are renamed to reflect the what
they are protecting.
Refs #10375