refs #10385. Naming issue root problem.
There is a naming issue here, which is actually the root cause of the problem! The query that was being performed retrieved the number of workspace entries, NOT the number of periods.
Going forward. If the number of periods exactly matches the number of workspace entries, we have a multiperiod group workspace, and can happily apply the optimisations we want.

Next step will be to actually extract the nperiods. To do that we need to try to fetch it out of the logs for The first entry.