Refs #10347 Use declarative bindings in options dialog
Instead of manually having to load/save options to/from widgets in
QtReflOptionsDialog, we now load/save options using a map binding
options to widgets. It may seem overkill for just one option, but when
we have fifteen to twenty it'll greatly simplify our life.

It could be expanded in future to build the bindings from widget
properties in the UI file, so new options need not even touch
QtReflOptionsDialog's code.