Refs #10090. Changed PoldiTruncateData.
As indicated in the comment in PoldiTruncateData, the chopper loading should have been replaced immediately after #9445, but since that did not cause problems with old data files, it was not necessary. Now it uses the code in PoldiInstrumentAdapter and PoldiTruncateData also works with new data files where the chopper speed is off.